+++ site/dat/docs/Changelog.md
@@ -4,6 +4,7 @@
  - add `cloud` and `local` aliases 
  - fix delay trouble in provisioning
  - fix Gatling keepalive type cast bug
+ - add `detach` mode to Cloud Provisioning
 
 ## 1.6.6 <sup>08 aug 2016</sup>
  - optimize aggregator by removing excessive calls to `BetterDict.get()`